requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
av==16.0.1
beautifulsoup4==4.14.2
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
decord==0.6.0
deep-translator==1.11.4
docopt==0.6.2
einops==0.8.1
fastapi==0.122.0
ffmpy==1.0.0
filelock==3.19.1
fsspec==2025.9.0
gradio==6.0.0
gradio_client==2.0.0.dev3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.5
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
num2words==0.5.14
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
psutil==7.1.3
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-triton==3.5.1+gitbfeb0668
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safehttpx==0.1.7
safetensors==0.7.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.10.0.dev20251122+cu128
torchaudio==2.10.0.dev20251122+cu128
torchvision==0.25.0.dev20251122+cu128
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers.git@5169c2364ba1ad2f555b83eac1b9065aa2955101
triton==3.0.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
PyYAML